src/test/java/com/namhla/Design_Patterns/test/SingletonTest.java
/*
 * To change this license header, choose License Headers in Project Properties.
 * To change this template file, choose Tools | Templates
 * and open the template in the editor.
 */

package com.namhla.Design_Patterns.test;

import com.namhla.design_patterns.creational.singleton.Singleton;
import static org.testng.Assert.*;

/**
 *
 * @author namhla
 */
public class SingletonTest {
    
    public SingletonTest() {
    }

    // TODO add test methods here.
    // The methods must be annotated with annotation @Test. For example:
    //
    // @Test
     public void SingletonTest() {
    String hello = null;
        Singleton a =  Singleton.instance();
        
        hello = a.sayHello();
        
        assertEquals(hello, "Hello");
}

    @org.testng.annotations.BeforeClass
    public static void setUpClass() throws Exception {
    }

    @org.testng.annotations.AfterClass
    public static void tearDownClass() throws Exception {
    }

    @org.testng.annotations.BeforeMethod
    public void setUpMethod() throws Exception {
    }

    @org.testng.annotations.AfterMethod
    public void tearDownMethod() throws Exception {
    }

    private void assertEquals(String hello, String hello0) {
        throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
    }
}
